priors/requirements.txt
Flask
google-generativeai
gunicorn
